Refactoring preferences part
