+++ TweetNestKit/ManagedObjects/ManagedPreferences+CoreDataProperties.swift
